splits domain and service commands
