+++ openvair/modules/template/adapters/dto/internal/commands/service.py
@@ -1,10 +1,9 @@
 from uuid import UUID  # noqa: D100
 from typing import ClassVar, Optional
-from pathlib import Path
 
 from pydantic import Field, BaseModel, ConfigDict
 
-from openvair.common.configs.pydantic import dto_config, lenient_dto_config
+from openvair.common.configs.pydantic import dto_config
 
 
 class CreateTemplateServiceCommandDTO(BaseModel):  # noqa: D101
@@ -28,17 +27,4 @@ class EditTemplateServiceCommandDTO(BaseModel):  # noqa: D101
 
 
 class DeleteTemplateServiceCommandDTO(BaseModel):  # noqa: D101
-    id: UUID
-
-
-class CreateTemplateDomainCommandDTO(BaseModel):  # noqa: D101
-    source_disk_path: Path
-    model_config: ClassVar[ConfigDict] = lenient_dto_config
-
-
-class EditTemplateDomainCommandDTO(BaseModel):  # noqa: D101
-    name: str
-    description: Optional[str]
-
-class DeleteTemplateDomainCommandDTO(BaseModel):  # noqa: D101
     id: UUID
